Create Hardware namespace for components
