+++ RedBookPlayer/Hardware/Player.cs
@@ -1,17 +1,13 @@
 using System;
 using System.IO;
 using System.Linq;
-using System.Threading.Tasks;
 using Aaru.CommonTypes.Enums;
 using Aaru.DiscImages;
 using Aaru.Filters;
-using CSCore.SoundOut;
-using NWaves.Audio;
-using NWaves.Filters.BiQuad;
 using RedBookPlayer.Discs;
 using RedBookPlayer.GUI;
 
-namespace RedBookPlayer
+namespace RedBookPlayer.Hardware
 {
     public class Player
     {
@@ -23,14 +19,9 @@ namespace RedBookPlayer
         public bool Initialized { get; private set; } = false;
 
         /// <summary>
-        /// Indicates if de-emphasis should be applied
-        /// </summary>
-        public bool ApplyDeEmphasis { get; private set; } = false;
-
-        /// <summary>
         /// Indicate if the disc is playing
         /// </summary>
-        public bool Playing => _soundOut.PlaybackState == PlaybackState.Playing;
+        public bool Playing => _soundOutput?.Playing ?? false;
 
         #endregion
 
@@ -42,34 +33,9 @@ namespace RedBookPlayer
         private OpticalDisc _opticalDisc;
 
         /// <summary>
-        /// Current position in the sector
+        /// Sound output handling class
         /// </summary>
-        private int _currentSectorReadPosition = 0;
-
-        /// <summary>
-        /// Data provider for sound output
-        /// </summary>
-        private PlayerSource _source;
-
-        /// <summary>
-        /// Sound output instance
-        /// </summary>
-        private ALSoundOut _soundOut;
-
-        /// <summary>
-        /// Left channel de-emphasis filter
-        /// </summary>
-        private BiQuadFilter _deEmphasisFilterLeft;
-
-        /// <summary>
-        /// Right channel de-emphasis filter
-        /// </summary>
-        private BiQuadFilter _deEmphasisFilterRight;
-
-        /// <summary>
-        /// Lock object for reading track data
-        /// </summary>
-        private readonly object _readingImage = new object();
+        public SoundOutput _soundOutput;
 
         #endregion
 
@@ -82,7 +48,8 @@ namespace RedBookPlayer
         {
             // Reset the internal state for initialization
             Initialized = false;
-            ApplyDeEmphasis = false;
+            _soundOutput = new SoundOutput();
+            _soundOutput.ApplyDeEmphasis = false;
             _opticalDisc = null;
 
             try
@@ -106,146 +73,13 @@ namespace RedBookPlayer
                 return;
             }
 
-            // If we have an unusable disc, just return
-            if(_opticalDisc == null || !_opticalDisc.Initialized)
+            // Initialize the sound output
+            _soundOutput.Init(_opticalDisc, autoPlay);
+            if(_soundOutput == null || !_soundOutput.Initialized)
                 return;
-
-            // Enable de-emphasis for CDs, if necessary
-            if(_opticalDisc is CompactDisc compactDisc)
-                ApplyDeEmphasis = compactDisc.TrackHasEmphasis;
-
-            // Setup de-emphasis filters
-            SetupFilters();
-
-            // Setup the audio output
-            SetupAudio();
-
-            // Initialize playback, if necessary
-            if(autoPlay)
-                _soundOut.Play();
 
             // Mark the player as ready
             Initialized = true;
-
-            // Begin loading data
-            _source.Start();
-        }
-
-        /// <summary>
-        /// Fill the current byte buffer with playable data
-        /// </summary>
-        /// <param name="buffer">Buffer to load data into</param>
-        /// <param name="offset">Offset in the buffer to load at</param>
-        /// <param name="count">Number of bytes to load</param>
-        /// <returns>Number of bytes read</returns>
-        public int ProviderRead(byte[] buffer, int offset, int count)
-        {
-            // Set the current volume
-            _soundOut.Volume = (float)App.Settings.Volume / 100;
-
-            // Determine how many sectors we can read
-            ulong sectorsToRead;
-            ulong zeroSectorsAmount;
-            do
-            {
-                // Attempt to read 2 more sectors than requested
-                sectorsToRead = ((ulong)count / (ulong)_opticalDisc.BytesPerSector) + 2;
-                zeroSectorsAmount = 0;
-
-                // Avoid overreads by padding with 0-byte data at the end
-                if(_opticalDisc.CurrentSector + sectorsToRead > _opticalDisc.TotalSectors)
-                {
-                    ulong oldSectorsToRead = sectorsToRead;
-                    sectorsToRead = _opticalDisc.TotalSectors - _opticalDisc.CurrentSector;
-
-                    int tempZeroSectorCount = (int)(oldSectorsToRead - sectorsToRead);
-                    zeroSectorsAmount = (ulong)(tempZeroSectorCount < 0 ? 0 : tempZeroSectorCount);
-                }
-
-                // TODO: Figure out when this value could be negative
-                if(sectorsToRead <= 0)
-                {
-                    _opticalDisc.LoadFirstTrack();
-                    _currentSectorReadPosition = 0;
-                }
-            } while(sectorsToRead <= 0);
-
-            // Create padding data for overreads
-            byte[] zeroSectors = new byte[(int)zeroSectorsAmount * _opticalDisc.BytesPerSector];
-            byte[] audioData;
-
-            // Attempt to read the required number of sectors
-            var readSectorTask = Task.Run(() =>
-            {
-                lock(_readingImage)
-                {
-                    for (int i = 0; i < 4; i++)
-                    {
-                        try
-                        {
-                            return _opticalDisc.ReadSectors((uint)sectorsToRead).Concat(zeroSectors).ToArray();
-                        }
-                        catch(ArgumentOutOfRangeException)
-                        {
-                            _opticalDisc.LoadFirstTrack();
-                        }
-                    }
-
-                    return zeroSectors;
-                }
-            });
-
-            // Wait 100ms at longest for the read to occur
-            if(readSectorTask.Wait(TimeSpan.FromMilliseconds(100)))
-            {
-                audioData = readSectorTask.Result;
-            }
-            else
-            {
-                Array.Clear(buffer, offset, count);
-                return count;
-            }
-
-            // Load only the requested audio segment
-            byte[] audioDataSegment = new byte[count];
-            int copyAmount = Math.Min(count, audioData.Length - _currentSectorReadPosition);
-            if(Math.Max(0, copyAmount) == 0)
-            {
-                Array.Clear(buffer, offset, count);
-                return count;
-            }
-
-            Array.Copy(audioData, _currentSectorReadPosition, audioDataSegment, 0, copyAmount);
-
-            // Apply de-emphasis filtering, only if enabled
-            if(ApplyDeEmphasis)
-            {
-                float[][] floatAudioData = new float[2][];
-                floatAudioData[0] = new float[audioDataSegment.Length / 4];
-                floatAudioData[1] = new float[audioDataSegment.Length / 4];
-                ByteConverter.ToFloats16Bit(audioDataSegment, floatAudioData);
-
-                for(int i = 0; i < floatAudioData[0].Length; i++)
-                {
-                    floatAudioData[0][i] = _deEmphasisFilterLeft.Process(floatAudioData[0][i]);
-                    floatAudioData[1][i] = _deEmphasisFilterRight.Process(floatAudioData[1][i]);
-                }
-
-                ByteConverter.FromFloats16Bit(floatAudioData, audioDataSegment);
-            }
-
-            // Write out the audio data to the buffer
-            Array.Copy(audioDataSegment, 0, buffer, offset, count);
-
-            // Set the read position in the sector for easier access
-            _currentSectorReadPosition += count;
-            if(_currentSectorReadPosition >= _opticalDisc.BytesPerSector)
-            {
-                _opticalDisc.CurrentSector += (ulong)(_currentSectorReadPosition / _opticalDisc.BytesPerSector);
-                _currentSectorReadPosition %= _opticalDisc.BytesPerSector;
-            }
-
-            return count;
         }
 
         #region Playback
@@ -261,12 +95,12 @@ namespace RedBookPlayer
 
             if(start)
             {
-                _soundOut.Play();
+                _soundOutput.Play();
                 _opticalDisc.SetTotalIndexes();
             }
             else
             {
-                _soundOut.Stop();
+                _soundOutput.Stop();
             }
         }
 
@@ -278,7 +112,7 @@ namespace RedBookPlayer
             if(_opticalDisc == null || !_opticalDisc.Initialized)
                 return;
 
-            _soundOut.Stop();
+            _soundOutput.Stop();
             _opticalDisc.LoadFirstTrack();
         }
 
@@ -295,7 +129,7 @@ namespace RedBookPlayer
 
             _opticalDisc.NextTrack();
             if(_opticalDisc is CompactDisc compactDisc)
-                ApplyDeEmphasis = compactDisc.TrackHasEmphasis;
+                _soundOutput.ApplyDeEmphasis = compactDisc.TrackHasEmphasis;
 
             if(wasPlaying) TogglePlayPause(true);
         }
@@ -313,7 +147,7 @@ namespace RedBookPlayer
 
             _opticalDisc.PreviousTrack();
             if(_opticalDisc is CompactDisc compactDisc)
-                ApplyDeEmphasis = compactDisc.TrackHasEmphasis;
+                _soundOutput.ApplyDeEmphasis = compactDisc.TrackHasEmphasis;
 
             if(wasPlaying) TogglePlayPause(true);
         }
@@ -332,7 +166,7 @@ namespace RedBookPlayer
 
             _opticalDisc.NextIndex(changeTrack);
             if(_opticalDisc is CompactDisc compactDisc)
-                ApplyDeEmphasis = compactDisc.TrackHasEmphasis;
+                _soundOutput.ApplyDeEmphasis = compactDisc.TrackHasEmphasis;
 
             if(wasPlaying) TogglePlayPause(true);
         }
@@ -351,7 +185,7 @@ namespace RedBookPlayer
 
             _opticalDisc.PreviousIndex(changeTrack);
             if(_opticalDisc is CompactDisc compactDisc)
-                ApplyDeEmphasis = compactDisc.TrackHasEmphasis;
+                _soundOutput.ApplyDeEmphasis = compactDisc.TrackHasEmphasis;
 
             if(wasPlaying) TogglePlayPause(true);
         }
@@ -424,7 +258,7 @@ namespace RedBookPlayer
         /// Toggle de-emphasis processing
         /// </summary>
         /// <param name="enable">True to apply de-emphasis, false otherwise</param>
-        public void ToggleDeEmphasis(bool enable) => ApplyDeEmphasis = enable;
+        public void ToggleDeEmphasis(bool enable) => _soundOutput.ToggleDeEmphasis(enable);
 
         /// <summary>
         /// Update the data context for the frontend
@@ -436,7 +270,7 @@ namespace RedBookPlayer
                 return;
 
             dataContext.HiddenTrack = _opticalDisc.TimeOffset > 150;
-            dataContext.ApplyDeEmphasis = ApplyDeEmphasis;
+            dataContext.ApplyDeEmphasis = _soundOutput.ApplyDeEmphasis;
 
             if(_opticalDisc is CompactDisc compactDisc)
             {
@@ -451,40 +285,6 @@ namespace RedBookPlayer
                 dataContext.IsDataTrack = _opticalDisc.TrackType != TrackType.Audio;
                 dataContext.CopyAllowed = false;
                 dataContext.TrackHasEmphasis = false;
-            }
-        }
-
-        /// <summary>
-        /// Sets or resets the de-emphasis filters
-        /// </summary>
-        private void SetupFilters()
-        {
-            if(_deEmphasisFilterLeft == null)
-            {
-                _deEmphasisFilterLeft = new DeEmphasisFilter();
-                _deEmphasisFilterRight = new DeEmphasisFilter();
-            }
-            else
-            {
-                _deEmphasisFilterLeft.Reset();
-                _deEmphasisFilterRight.Reset();
-            }
-        }
-
-        /// <summary>
-        /// Sets or resets the audio playback objects
-        /// </summary>
-        private void SetupAudio()
-        {
-            if(_source == null)
-            {
-                _source = new PlayerSource(ProviderRead);
-                _soundOut = new ALSoundOut(100);
-                _soundOut.Initialize(_source);
-            }
-            else
-            {
-                _soundOut.Stop();
             }
         }
