+++ RedBookPlayer/Hardware/PlayerSource.cs
@@ -2,7 +2,7 @@ using System;
 using CSCore;
 using WaveFormat = CSCore.WaveFormat;
 
-namespace RedBookPlayer
+namespace RedBookPlayer.Hardware
 {
     public class PlayerSource : IWaveSource
     {
